fix(warnings): fix build warnings and errors on macOS and Windows CI
- mrpt_system/crc.cpp: explicit cast int→uint16_t (-Wconversion)
- mrpt_system/md5.cpp: sprintf→snprintf (-Wdeprecated-declarations)
- mrpt_system/memory.cpp: old-style cast → reinterpret_cast (-Wold-style-cast)
- mrpt_math/TPoint2D.h: static_cast<double> to suppress -Wdouble-promotion
- mrpt_poses: (void)normalizeWeights() to suppress -Wunused-result
- mrpt_viz/CPointCloudColoured.h: remove unused private field
- mrpt_graphs/CGraphPartitioner.cpp: float→bool via explicit == 0 check
- mrpt_obs/CObservation3DRangeScan_project3D_impl.h: explicit float casts for
  camera params; [[maybe_unused]] on SSE2-only function parameters
- mrpt_obs/CObservationGasSensors.cpp: sprintf→snprintf
- mrpt_obs/VelodyneCalibration.cpp: remove unused loop variable
- mrpt_opengl/CompiledViewport.h: [[maybe_unused]] on unused private field
- mrpt_imgui/CImGuiSceneView.cpp: define GL_SILENCE_DEPRECATION on Apple
- mrpt_gui/about_box_*.cpp: (void) casts for params unused in #else branches
- mrpt_gui/CDisplayWindow3D.cpp: (void) casts for params/captures unused in
  #else branches
- mrpt_gui/CDisplayWindowPlots.cpp: guard mathplot.h include with
  #if MRPT_HAS_WXWIDGETS (macOS fatal error: wx/dcmemory.h not found)
- mrpt_containers/cmake/find_libfyaml.cmake: skip built-in libfyaml on Windows
  (requires POSIX pthreads + sys/mman.h unavailable on MSVC)

+++ modules/mrpt_system/src/crc.cpp
@@ -51,7 +51,7 @@ uint16_t mrpt::system::compute_CRC16(const uint8_t* data, size_t len_, const uin
     {
       uCrc16 <<= 1;
     }
-    uCrc16 ^= (abData[0] | (abData[1] << 8));
+    uCrc16 ^= static_cast<uint16_t>(abData[0] | (abData[1] << 8));
   }
   return uCrc16;
 }
